requirements.txt
# Monday BI Agent Dependencies
streamlit
pandas
requests
python-dotenv
openai
